% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-observation.R
\name{geom_observation}
\alias{geom_observation}
\title{Observation}
\usage{
geom_observation(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...,
  prefix = "n = ",
  suffix = "",
  compute = "group",
  separation_factor = 1
)
}
\arguments{
\item{computer}{by group or by panel}
}
\description{
Adds the number of observations
}
